+++ public/index.html
@@ -1567,6 +1567,6 @@ This is card with missing deadline</textarea
       </section>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="bundle.js"></script>
   </body>
 </html>
