+++ src/data.js
@@ -1,0 +1,33 @@
+// Список фильтров
+const FILTER_NAME_LIST = [
+  `all`,
+  `overdue`,
+  `today`,
+  `favorites`,
+  `repeating`,
+  `tags`,
+  `archive`
+];
+
+// Список типов карточек
+const CARD_TYPE_LIST = [
+  ``,
+  `card--edit`
+];
+
+// Список типов баров у карточек
+const CARD_BAR_TYPE_LIST = [
+  ``,
+  `card--repeat`
+];
+
+// Список цветов баров у карточек
+const CARD_COLOR_LIST = [
+  `card--pink`,
+  `card--blue`,
+  `card--black`,
+  `card--yellow`,
+  `card--deadline`
+];
+
+export {FILTER_NAME_LIST, CARD_TYPE_LIST, CARD_BAR_TYPE_LIST, CARD_COLOR_LIST};
